pyproject.toml
[project]
name = "etl-tienda-mascotas"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "rich>=13.0",
    "pandas>=2.0",
    "openpyxl>=3.1",
    "google-auth>=2.0",
    "google-auth-oauthlib>=1.0",
    "google-api-python-client>=2.0",
    "requests>=2.31",
    "dash>=2.14",
    "plotly>=5.18",
    "dash-bootstrap-components>=1.5",
    "python-dotenv>=1.0",
    "pandas-stubs>=3.0.0.260204",
]

[tool.basedpyright]
# Mismo entorno que `uv sync` / desarrollo (pandas-stubs está en .venv)
venvPath = "."
venv = ".venv"
include = ["src"]
pythonVersion = "3.12"
reportMissingTypeStubs = "none"

[tool.pyright]
# Compatibilidad con Pyright clásico
venvPath = "."
venv = ".venv"
include = ["src"]
pythonVersion = "3.12"
reportMissingTypeStubs = "none"

[tool.ruff]
line-length = 79
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["config", "src"]
